+++ src/main/java/subway/station/StationService.java
@@ -1,4 +1,4 @@
-package subway;
+package subway.station;
 
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -26,6 +26,8 @@ public class StationService {
                 .map(this::createStationResponse)
                 .collect(Collectors.toList());
     }
+
+
 
     @Transactional
     public void deleteStationById(Long id) {
